+++ packages/http/src/forwarder/resources.ts
@@ -1,0 +1,10 @@
+export const hopByHopHeaders: string[] = [
+  'transfer-encoding',
+  'connection',
+  'keep-alive',
+  'proxy-authenticate',
+  'proxy-authorization',
+  'te',
+  'trailers',
+  'upgrade',
+];